% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcheck.R
\name{dfcheck}
\alias{dfcheck}
\title{Check whether an input is a data frame or not}
\usage{
dfcheck(x)
}
\arguments{
\item{x}{anything. An error is raised if it is not a data frame}
}
\value{
This function returns nothing. An error is raised if the input is not a data frame
}
\description{
Check if an input is a data frame or not. This is a internal function.
}
