% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFactors.R
\docType{data}
\name{exampleFactor1}
\alias{exampleFactor1}
\title{Example factor 1}
\format{A factor with 3 levels of length 3}
\usage{
exampleFactor1
}
\description{
Example factor 1
}
\details{
Alias for \code{iris$Species[c(1, 51, 101)]}
}
\keyword{datasets}

