% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFactors.R
\docType{data}
\name{exampleFactor2}
\alias{exampleFactor2}
\title{Example factor 2}
\format{A factor with 3 levels of length 3}
\usage{
exampleFactor2
}
\description{
Example factor 2
}
\details{
Alias for \code{PlantGrowth$group[c(1, 11, 21)]}
}
\keyword{datasets}

