% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFactors.R
\docType{data}
\name{exampleFactor3}
\alias{exampleFactor3}
\title{Example factor 3}
\format{A factor with 3 levels of length 34}
\usage{
exampleFactor3
}
\description{
Example factor 3
}
\details{
Alias for \code{iris$Species[c(1:10, 51:65, 101:109)]}
}
\keyword{datasets}

