% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors together}
\usage{
fbind(a, b, level.as.appears = TRUE)
}
\arguments{
\item{a}{factor}

\item{b}{factor}

\item{level.as.appears}{boolean; should the factor levels be combined as they appear?}
}
\value{
factor
}
\description{
Bind two factors together
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}

