% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcheck.R
\name{fcheck}
\alias{fcheck}
\title{Check whether an input is a factor or not}
\usage{
fcheck(x)
}
\arguments{
\item{x}{anything. An error is raised if it is not a factor.}
}
\value{
This function returns nothing. An error is raised if the input is not a factor.
}
\description{
Check if an input is a factor or not. This is a internal function.
}
