% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{Freqeqncy Table of factos levels}
\usage{
fcount(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Freqeqncy Table of factos levels
}
\examples{
fcount(iris$Species)
}
