% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdetect.R
\name{fdetect}
\alias{fdetect}
\title{Detect whether a factor should be character}
\usage{
fdetect(f)
}
\arguments{
\item{f}{a factor. An error is raised if it is not a factor.}
}
\value{
This function returns a boolean. It returns \code{TRUE} if the input is a factor,
\code{FALSE} if the input is a character. An error is raised if the input is not a factor.
}
\description{
Check if a factor should be character based on the fact that
a character should have a length equals to its unique values, while a factor should not.
}
\examples{
fdetect(factor(c("s", "t", "a", "t"))) # TRUE
fdetect(factor(c("s", "t", "a"))) # FALSE

}
