% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freorder.R
\name{freorder}
\alias{freorder}
\title{Reorder a factor in descending order}
\usage{
freorder(f)
}
\arguments{
\item{f}{a factor. An error is raised if it is not a factor.}
}
\value{
This function returns a factor in descending order.
An error is raised if the input is not a factor.
}
\description{
Reorder a factor in descending order.
}
\examples{
freorder(factor(c("s", "t", "a"))) # Levels: t s a

}
