% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_out.R
\name{freq_out}
\alias{freq_out}
\title{Create a frequency table for a factor}
\usage{
freq_out(x, n.sort = NULL)
}
\arguments{
\item{x}{factor}

\item{n.sort}{one of \code{c(NULL, "ascending", "descending")}. This argument specifies how to sort the counts in the frequency table.}
}
\value{
tbl_df
}
\description{
Create a frequency table for a factor
}
\examples{
freq_out(iris$Species)
}

