% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fset.R
\name{fset}
\alias{fset}
\title{Set the levels "as is"}
\usage{
fset(f)
}
\arguments{
\item{f}{a factor. An error is raised if it is not a factor.}
}
\value{
This function returns a factor, whose levels is set to the order of its data.
An error is raised if the input is not a factor.
}
\description{
Set levels of a factor to the order in which they appear in the data.
}
\examples{
fset(factor(c("s", "t", "a"))) # Levels: s t a

}
