% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should_be_factor.R
\name{should_be_factor}
\alias{should_be_factor}
\title{Should this really be a factor?}
\usage{
should_be_factor(x)
}
\arguments{
\item{x}{vector}
}
\value{
boolean; \code{TRUE} if the number of unique values does not equal to the length, and \code{FALSE} otherwise.
}
\description{
Should this really be a factor?
}
\examples{
# This should be a factor:
should_be_factor(c("a", "b", "b"))

# This should not be a factor
should_be_factor(c("a", "b", "c"))
}

