% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should_factor_stay_factor.R
\name{should_factor_stay_factor}
\alias{should_factor_stay_factor}
\title{Should this really be a factor?}
\usage{
should_factor_stay_factor(x)
}
\arguments{
\item{x}{factor}
}
\value{
boolean; \code{TRUE} if the number of unique values does not equal to the length, and \code{FALSE} otherwise.
}
\description{
Should this really be a factor?
}
\examples{
abb <- factor(c("a", "b", "b"))
abc <- factor(c("a", "b", "c"))

# This should be a factor:
should_factor_stay_factor(abb)

# This should not be a factor
should_factor_stay_factor(abc)
}

