% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfread.R, R/dfwrite.R
\name{dfread}
\alias{dfread}
\alias{dfwrite}
\title{Write and data frame}
\usage{
dfread(dffilename, lvfilename = NA)

dfwrite(df, dffilename, lvfilename = "levels.txt")
dfread(dffilename, lvfilename = "levels.txt")
}
\arguments{
\item{dffilename}{path to store the file of the data frame.}

\item{lvfilename}{(optional) path to store the companion file.
If it is not set, a "levels.txt" is used under the same folder of the file for data frame.}

\item{df}{a data frame. An error is raised if it is not a data frame.}
}
\value{
\code{fwrite()} returns nothing. \code{fread()} returns a data frame read from files.
An error is raised if the input is not a dataframe, or files not found.
Two files are generated: one text file for data frame, one companion file for factor levels.
}
\description{
Write and read data frame to/from plain text delimited files
while retaining factor levels.
}
\examples{
dfwrite(gapminder, "gapminder.txt", "levels.txt")
dfread("gapminder.txt", "levels.txt") # return gapminder data frame

}
